Extending and updating architecture to reflect input flow
As I began thinking through inputs and eventual support for rebinding (while also attempting to keep it as pure ECS as possible), I realized there was a lot more complexity than expected.  I have added a C3 diagram for the input system to reflect my thought process on this.  (I also adjusted the C2 to better reflect the new thinking.)