SDLEventSource file had incorrect casing for the current standard
I noticed as I added SdlRenderer, that this file did not follow the standard of acronyms per Google guidelines.  It's worth noting this will make it inconsistent with the library itself (e.g., SDLEvent), but I think it's important to try and maintain the standard insofar as possible.

Also, I would normally have pushed this change to main in a multi-developer environment, but just pushing it heree for my own ease.