Adding a SimpleRenderableShape component
For now, this component only supports rectangles and circles (possibly all that will be needed for the first pass at the game).

I also added a few future thinking collider components while it was in my brain to do so.  (I'd probably save these changes for future me in a normal project, but since I am already thinking about circles and rects, it seemed reasonable to change here and now.)