Fixes input loss by implementing a quit command properly
Having not worked with SDL much, I wasn't considering accidentally clearing the inputs by checking for an SDL_QUIT_EVENT external to the input polling (though it obviously makes sense.)  I was trying to defer additional commands to the game loop milestone, but per my comments, I knew it would be a more proper way top manage inputs, as well as provide an option to bind quit to a key press (ESC).

Technically, there should be a confirmation check on quit, but that will be a future polish item.

I tracked this as a "bug" though it did ultimately turn out to be a feature to implement.