Adding tests for new systems + removing GW from `DestructionSystem` constructor
I was going to attempt a refactor of `GameWorld` to make it much more test friendly, but I have determined this is a large undertaking.  For now, I'm just allowing these to be more "integration tests."

Also, I removed the need for a `GameWorld` to be injected into DestructionSystem` since I now pass it through the update.